pyproject.toml
[tool.poetry]
name = "resiclejeinteligente"
version = "1.0.0"
description = "esta es una app de resiclaje"
authors = ["Brandon <bg4710699@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10.9"
django-cors-headers = "^4.2.0"
django = "^4.2.5"
gunicorn = "^21.2.0"
whitenoise = {extras = ["brotli"], version = "^6.5.0"}
djangorestframework = "^3.14.0"
cloudinary = "^1.34.0"
dj-database-url = "^2.1.0"
psycopg2 = "^2.9.7"
drf-extra-fields = "^3.7.0"
pillow = "^10.0.1"
djangorestframework-simplejwt = "^5.3.0"


